+++ site/src/main.rs
@@ -269,7 +269,7 @@ fn prim_class(prim: Primitive) -> &'static str {
 
     match prim {
         Primitive::Transpose => code_font!("monadic-function-button trans"),
-        Primitive::Lives => code_font!("noadic-function-button"),
+        Primitive::Share => code_font!("noadic-function-button"),
         prim if prim.class() == PrimClass::Stack && prim.modifier_args().is_none() => {
             code_font!("stack-function-button")
         }