+++ src/algorithm/fork.rs
@@ -189,21 +189,21 @@ pub fn trident(env: &mut Uiua) -> UiuaResult {
     Ok(())
 }
 
-pub fn lives(env: &mut Uiua) -> UiuaResult {
+pub fn share(env: &mut Uiua) -> UiuaResult {
     let f = env.pop(FunctionArg(1))?;
     let g = env.pop(FunctionArg(2))?;
     let f = f
         .into_func_array()
         .map_err(|val| {
             env.error(format!(
-                "Lives' first function must be a function, but is has type {}",
+                "Share' first function must be a function, but is has type {}",
                 val.type_name()
             ))
         })
         .and_then(|f| {
             f.into_scalar().map_err(|arr| {
                 env.error(format!(
-                    "Lives' first function must be a scalar, but is has rank {}",
+                    "Share' first function must be a scalar, but is has rank {}",
                     arr.rank()
                 ))
             })
@@ -212,14 +212,14 @@ pub fn lives(env: &mut Uiua) -> UiuaResult {
         .into_func_array()
         .map_err(|val| {
             env.error(format!(
-                "Lives' second function must be a function, but is has type {}",
+                "Share' second function must be a function, but is has type {}",
                 val.type_name()
             ))
         })
         .and_then(|f| {
             f.into_scalar().map_err(|arr| {
                 env.error(format!(
-                    "Lives' second function must be a scalar, but is has rank {}",
+                    "Share' second function must be a scalar, but is has rank {}",
                     arr.rank()
                 ))
             })