+++ src/primitive/mod.rs
@@ -171,7 +171,7 @@ impl Primitive {
                 Some(format!("try using dip{} instead", Primitive::Dip))
             }
             Primitive::Fork | Primitive::Trident => {
-                Some(format!("try using lives{} instead", Primitive::Lives))
+                Some(format!("try using share{} instead", Primitive::Share))
             }
             _ => None,
         }
@@ -487,7 +487,7 @@ impl Primitive {
             Primitive::Both => fork::both(env)?,
             Primitive::Fork => fork::fork(env)?,
             Primitive::Trident => fork::trident(env)?,
-            Primitive::Lives => fork::lives(env)?,
+            Primitive::Share => fork::share(env)?,
             Primitive::If => fork::iff(env)?,
             Primitive::Try => {
                 let f = env.pop(FunctionArg(1))?;